pyproject.toml
[tool.poetry]
name = "google-sheets"
version = "0.1.0"
description = ""
authors = ["Alinocco <alinoco.dia@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "^2.56.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.2"
oauth2client = "^4.1.3"
apiclient = "^1.0.4"
Django = "^4.1"
psycopg2 = "^2.9.3"
django-environ = "^0.9.0"
celery = "^5.2.7"
redis = "^4.3.4"
beautifulsoup4 = "^4.11.1"
lxml = "^4.9.1"
aiogram = "^2.21"
notifiers = "^1.3.3"
matplotlib = "^3.5.3"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
